Add configuration for individual facebook datasets
This project will be running on Facebook pages that post other content, so the config expects a list of post ids for the relevant posts to download comments from. Posts are grouped into datasets so that we can collect all the posts on one topic/question into a single dataset in coda + analysis. Note that using post ids is a slow, laborious solution - we've asked M2A to include our project hashtag in all of our projects so we can have a means of automatically identifying our posts, but this PR post ids first as it is simpler and less dependent on M2A to remember our hashtag.

The design continues to mirror that of RapidProClient and RapidProSource, the main difference being the formatting of the keys, which uses '.' notation rather than the weird combination of casing, brackets, and hyphens that the rapid pro client uses.

Note that the configuration is still all placeholder, pointing to random imaqal posts for now.